Extract issuance protocol FSM
