+++ aries_vcx/src/handlers/issuance/issuer/mod.rs
@@ -5,8 +5,6 @@ use crate::messages::issuance::credential_proposal::CredentialProposal;
 
 pub mod issuer;
 pub mod utils;
-mod state_machine;
-mod states;
 
 pub async fn get_credential_proposal_messages(connection: &Connection) -> VcxResult<String> {
     let credential_proposals: Vec<CredentialProposal> = connection.get_messages()
