+++ aries_vcx/src/protocols/issuance/holder/mod.rs
@@ -1,0 +1,2 @@
+pub mod state_machine;
+pub mod states;
