+++ aries_vcx/src/protocols/issuance/holder/state_machine.rs
@@ -1,24 +1,24 @@
 use std::collections::HashMap;
 
 use crate::error::prelude::*;
-use crate::handlers::SendClosure;
 use crate::handlers::issuance::holder::holder::HolderState;
-use crate::handlers::issuance::holder::states::finished::FinishedHolderState;
-use crate::handlers::issuance::holder::states::offer_received::OfferReceivedState;
-use crate::handlers::issuance::holder::states::request_sent::RequestSentState;
-use crate::handlers::issuance::holder::states::proposal_sent::ProposalSentState;
-use crate::handlers::issuance::holder::states::initial::InitialHolderState;
 use crate::handlers::issuance::messages::CredentialIssuanceMessage;
 use crate::handlers::issuance::verify_thread_id;
+use crate::handlers::SendClosure;
 use crate::libindy::utils::anoncreds::{self, get_cred_def_json, libindy_prover_create_credential_req, libindy_prover_delete_credential, libindy_prover_store_credential};
-use crate::messages::a2a::{MessageId, A2AMessage};
+use crate::messages::a2a::{A2AMessage, MessageId};
 use crate::messages::error::ProblemReport;
 use crate::messages::issuance::credential::Credential;
 use crate::messages::issuance::credential_ack::CredentialAck;
 use crate::messages::issuance::credential_offer::CredentialOffer;
-use crate::messages::issuance::credential_request::CredentialRequest;
 use crate::messages::issuance::credential_proposal::CredentialProposal;
+use crate::messages::issuance::credential_request::CredentialRequest;
 use crate::messages::status::Status;
+use crate::protocols::issuance::holder::states::finished::FinishedHolderState;
+use crate::protocols::issuance::holder::states::initial::InitialHolderState;
+use crate::protocols::issuance::holder::states::offer_received::OfferReceivedState;
+use crate::protocols::issuance::holder::states::proposal_sent::ProposalSentState;
+use crate::protocols::issuance::holder::states::request_sent::RequestSentState;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub enum HolderFullState {
@@ -138,7 +138,7 @@ impl HolderSM {
                         VcxError::from_msg(VcxErrorKind::InvalidState, "Attempted to call undefined send_message callback")
                     )?(proposal.to_a2a_message()).await?;
                     HolderFullState::ProposalSent(ProposalSentState::new(proposal))
-                },
+                }
                 _ => {
                     HolderFullState::Initial(state_data)
                 }
@@ -146,7 +146,7 @@ impl HolderSM {
             HolderFullState::ProposalSent(state_data) => match cim {
                 CredentialIssuanceMessage::CredentialOffer(offer) => {
                     HolderFullState::OfferReceived(OfferReceivedState::new(offer))
-                },
+                }
                 CredentialIssuanceMessage::ProblemReport(problem_report) => {
                     HolderFullState::Finished(problem_report.into())
                 }
@@ -185,15 +185,15 @@ impl HolderSM {
                         VcxError::from_msg(VcxErrorKind::InvalidState, "Attempted to call undefined send_message callback")
                     )?(proposal.to_a2a_message()).await?;
                     HolderFullState::ProposalSent(ProposalSentState::new(proposal))
-                },
+                }
                 CredentialIssuanceMessage::CredentialOfferReject(comment) => {
-                        let problem_report = ProblemReport::create()
-                            .set_thread_id(&thread_id)
-                            .set_comment(comment);
-                        send_message.ok_or(
-                            VcxError::from_msg(VcxErrorKind::InvalidState, "Attempted to call undefined send_message callback")
-                        )?(problem_report.to_a2a_message()).await?;
-                        HolderFullState::Finished(problem_report.into())
+                    let problem_report = ProblemReport::create()
+                        .set_thread_id(&thread_id)
+                        .set_comment(comment);
+                    send_message.ok_or(
+                        VcxError::from_msg(VcxErrorKind::InvalidState, "Attempted to call undefined send_message callback")
+                    )?(problem_report.to_a2a_message()).await?;
+                    HolderFullState::Finished(problem_report.into())
                 }
                 _ => {
                     warn!("Unable to process received message in this state");
