+++ aries_vcx/src/protocols/issuance/holder/states/proposal_sent.rs
@@ -1,14 +1,14 @@
 use crate::error::prelude::*;
-use crate::messages::issuance::credential_proposal::CredentialProposal;
 use crate::handlers::issuance::is_cred_def_revokable;
+use crate::messages::issuance::credential_proposal::CredentialProposal;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct ProposalSentState {
-    pub credential_proposal: CredentialProposal
+    pub credential_proposal: CredentialProposal,
 }
 
 impl ProposalSentState {
-    pub fn new (credential_proposal: CredentialProposal) -> Self {
+    pub fn new(credential_proposal: CredentialProposal) -> Self {
         Self { credential_proposal }
     }
