+++ aries_vcx/src/protocols/issuance/issuer/states/finished.rs
@@ -1,4 +1,4 @@
-use crate::handlers::issuance::issuer::state_machine::RevocationInfoV1;
+use crate::protocols::issuance::issuer::state_machine::RevocationInfoV1;
 use crate::messages::status::Status;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
