+++ aries_vcx/src/protocols/issuance/issuer/states/initial.rs
@@ -1,4 +1,4 @@
-use crate::handlers::issuance::issuer::states::offer_sent::OfferSentState;
+use crate::protocols::issuance::issuer::states::offer_sent::OfferSentState;
 use crate::messages::issuance::credential_offer::{CredentialOffer, OfferInfo};
 
 #[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq)]
