+++ aries_vcx/src/protocols/issuance/issuer/states/offer_sent.rs
@@ -1,6 +1,6 @@
-use crate::handlers::issuance::issuer::state_machine::RevocationInfoV1;
-use crate::handlers::issuance::issuer::states::finished::FinishedState;
-use crate::handlers::issuance::issuer::states::requested_received::RequestReceivedState;
+use crate::protocols::issuance::issuer::state_machine::RevocationInfoV1;
+use crate::protocols::issuance::issuer::states::finished::FinishedState;
+use crate::protocols::issuance::issuer::states::requested_received::RequestReceivedState;
 use crate::messages::error::ProblemReport;
 use crate::messages::issuance::credential_offer::CredentialOffer;
 use crate::messages::issuance::credential_request::CredentialRequest;
